The base line game. There are still lots of features that need to be implemented
I have the maps, enemies, spawners, attractors, guns, bullets, damage, and no teleport zones working. It was lots of work, but it's going pretty well. I still need the enemies to do damage, a collider for them to do damage to, health for the tower, a timer for each level, and an upgrade system